+++ CHANGELOG.md
@@ -68,6 +68,7 @@ instead of use of M2O bankOrder of InvoicePayment object. Like that, if we remov
 - Partner : Do not clean contacts on partner type change but clean on save
 - Partner : Resolve error on save when modify fields and contactSet at same time. 
 - Interco: Compute price discounted for sale and purchase orders.
+- Generating a reversion of a stock move now correctly updates delivered/received quantity in linked order.
 
 ## [5.1.0] - 2018-12-13
 ## Features